+++ test/user_agent.exs
@@ -4,7 +4,11 @@ defmodule UserAgentTest do
   import UserAgent
   doctest UserAgent
 
-  test "Need add test" do
-    assert true == false
+  test "Test random" do
+    assert is_binary(UserAgent.random()) == true
+  end
+
+  test "Test random_all" do
+    assert is_binary(UserAgent.random_all()) == true
   end
 end