+++ CHANGELOG.md
@@ -1,2 +1,42 @@
-# Changelog
+# [1.0.0-alpha.2](https://github.com/gsmlg-dev/user_agent_generator/compare/v1.0.0-alpha.1...v1.0.0-alpha.2) (2022-08-16)
 
+
+### Bug Fixes
+
+* Fix description. ([05bc3fd](https://github.com/gsmlg-dev/user_agent_generator/commit/05bc3fd6f12741fb75c88add7233872921d75437))
+* Fix package. ([eef915e](https://github.com/gsmlg-dev/user_agent_generator/commit/eef915e1e64b2460843f94f2bd656879ab5052bb))
+
+# 1.0.0-alpha.1 (2022-08-16)
+
+
+### Bug Fixes
+
+* Fix deps. ([42364b2](https://github.com/gsmlg-dev/user_agent_generator/commit/42364b2e1ae7a5f27c9c3969c08e8a5115d50eb6))
+* Fix missing task. ([f268c8f](https://github.com/gsmlg-dev/user_agent_generator/commit/f268c8ff42f99a34cb59375dc008aa9ce758ce58))
+* Fix test. ([2ba2c11](https://github.com/gsmlg-dev/user_agent_generator/commit/2ba2c11c98ba0a949853c425eae239888dd22ad7))
+* Init package. ([2f84d86](https://github.com/gsmlg-dev/user_agent_generator/commit/2f84d86b0b2a440bcfd7e905dcf9ab3232062650))
+
+
+### Features
+
+* Add all_agents. ([42fc6dc](https://github.com/gsmlg-dev/user_agent_generator/commit/42fc6dc5f95c4f52e1cb3c57538c51aedfb79c5d))
+* Add most common user agents. ([d8d2f81](https://github.com/gsmlg-dev/user_agent_generator/commit/d8d2f81abfa97d827ba2b915245fa3b5e620c6f9))
+* Add random function. ([f65e5b4](https://github.com/gsmlg-dev/user_agent_generator/commit/f65e5b4f454fe18c19baaf66f0ce84e362377470))
+
+# 1.0.0-alpha.1 (2022-08-16)
+
+
+### Bug Fixes
+
+* Fix deps. ([42364b2](https://github.com/gsmlg-dev/user_agent_generator/commit/42364b2e1ae7a5f27c9c3969c08e8a5115d50eb6))
+* Fix test. ([2ba2c11](https://github.com/gsmlg-dev/user_agent_generator/commit/2ba2c11c98ba0a949853c425eae239888dd22ad7))
+* Init package. ([2f84d86](https://github.com/gsmlg-dev/user_agent_generator/commit/2f84d86b0b2a440bcfd7e905dcf9ab3232062650))
+
+
+### Features
+
+* Add all_agents. ([42fc6dc](https://github.com/gsmlg-dev/user_agent_generator/commit/42fc6dc5f95c4f52e1cb3c57538c51aedfb79c5d))
+* Add most common user agents. ([d8d2f81](https://github.com/gsmlg-dev/user_agent_generator/commit/d8d2f81abfa97d827ba2b915245fa3b5e620c6f9))
+* Add random function. ([f65e5b4](https://github.com/gsmlg-dev/user_agent_generator/commit/f65e5b4f454fe18c19baaf66f0ce84e362377470))
+
+# Changelog